Don't replace environment variables in the blueprint but do it at runtime
